Move interfaces from chat_stage.ts to new agent.ts
